LPS-199599 Inline CommercePunchOutThemeHttpHelper into CommercePunchOutThemeTemplateContextContributor
